# Conflicts:
#	front/src/BarreNav/BarreNavCore.js
#	front/src/Quiz/CreateProject.js
#	front/src/index.js
#	sql/init.sql

+++ back/src/main/java/com/ecoassitant/back/repository/ProfilRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 /**
  * repository of Profil
  */
-public interface ProfilRepository extends JpaRepository<ProfilEntity, Long> {
+public interface ProfilRepository extends JpaRepository<ProfilEntity, Integer> {
 
 
     /**
